server/requirements.txt
aniso8601==9.0.1
backports.zoneinfo==0.2.1
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mongoengine==0.26.0
openpyxl==3.1.0
PyJWT==2.6.0
pymongo==4.3.3
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
tendo==0.3.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uWSGI==2.0.21
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.12.0
shapely==2.0.1
Flask-Limiter==3.3.0
PyYAML==6.0.1
jsonschema==4.17.3
python-dateutil==2.8.2